+++ CHANGELOG.md
@@ -17,7 +17,10 @@ v2.0.0-beta.10
  * ons-alert-dialog: Fixed [#1347](https://github.com/OnsenUI/OnsenUI/issues/1347).
  * ons-alert-dialog, ons-dialog, ons-popover, ons-modal: Updated API (`disabled`, `visible`, `backButtonHandler` properties).
  * ons-progress-bar, ons-progress-circular: Add `value`, `secondaryValue` and `indeterminate` properties.
- * ons-pull-hook: Add `height`, `thresholdHeight` and `disabled` properties.
+ * ons-pull-hook: Add `state`, `pullDistance`, `height`, `thresholdHeight` and `disabled` properties.
+ * ons-page: Update API (`disabled`, `backButtonHandler` properties).
+ * ons-button, ons-ripple: Update API (`disabled` property).
+ * ons-speeddial: Updated API (`disabled`, `visible`, `inline` properties).
 
 v2.0.0-beta.9
 ----